refactor: Move useChartState to an appropriate file
