+++ app/charts/shared/axis-width-band.tsx
@@ -2,7 +2,7 @@ import { axisBottom, select, Selection } from "d3";
 import { useEffect, useRef } from "react";
 
 import { ColumnsState } from "@/charts/column/columns-state";
-import { useChartState } from "@/charts/shared/use-chart-state";
+import { useChartState } from "@/charts/shared/chart-state";
 import { useChartTheme } from "@/charts/shared/use-chart-theme";
 import { useTimeFormatUnit } from "@/formatters";
