+++ app/charts/shared/interaction/hover-dot.tsx
@@ -2,7 +2,7 @@ import { Box } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 
 import { LinesState } from "@/charts/line/lines-state";
-import { useChartState } from "@/charts/shared/use-chart-state";
+import { useChartState } from "@/charts/shared/chart-state";
 import { useInteraction } from "@/charts/shared/use-interaction";
 import { Observation } from "@/domain/data";
